AVReader: fix seek
* Do not seek when reading the first frame.
* Update message printed at each frame (tmp for debug).

+++ plugins/image/io/AudioVideo/src/reader/AVReaderPlugin.hpp
@@ -62,7 +62,7 @@ public:
 	
 	std::string _lastInputFilePath;
 	size_t _idVideoStream;
-	size_t _lastFrame;
+	int _lastFrame;
 };
 
 }
